requirements.txt
Flask==2.2.0
Flask-JWT-Extended==4.4.0
flasgger==0.9.5
python-dotenv==0.21.0